requirements.txt
python-telegram-bot
sqlalchemy
pymysql
python-dotenv